Para probar la raspi

+++ level.c
@@ -318,16 +318,15 @@ void drawLevel(estadoJuego_t* gameState){
             posY = (int) (gameState->entidades.enemigos[i].fisica.posy / PIXELSPERUNIT);
             mapLevel[posY][posX] = 1;
 
-            /*
-            for (int j = 0; j < gameState->entidades.enemigos[i].fisica.ancho / PIXELSPERUNIT - 1; j++) {
+
+            for (int j = 0; j < ( (int) (gameState->entidades.enemigos[i].fisica.ancho / PIXELSPERUNIT) ) - 1; j++) {
                 mapLevel[posY][posX + j] = 1;
             }
 
-            for (int j = 0; j < gameState->entidades.enemigos[i].fisica.alto / PIXELSPERUNIT - 1; j++) {
+            for (int j = 0; j < ( (int) (gameState->entidades.enemigos[i].fisica.alto / PIXELSPERUNIT) ) - 1; j++) {
                 mapLevel[posY + j][posX] = 1;
             }
 
-            */
         }
     }
 
@@ -338,17 +337,14 @@ void drawLevel(estadoJuego_t* gameState){
             posY = (int) (gameState->entidades.bloques[i].fisica.posy / PIXELSPERUNIT);
             mapLevel[posY][posX] = 1;
 
-            /*
-
-            for (int j = 0; j < gameState->entidades.bloques[i].fisica.ancho / PIXELSPERUNIT - 1; j++) {
+            for (int j = 0; j < ( (int) (gameState->entidades.bloques[i].fisica.ancho) ) / PIXELSPERUNIT - 1; j++) {
                 mapLevel[posY][posX + j] = 1;
             }
 
-            for (int j = 0; j < gameState->entidades.bloques[i].fisica.alto / PIXELSPERUNIT - 1; j++) {
+            for (int j = 0; j < ( (int) (gameState->entidades.bloques[i].fisica.alto) ) / PIXELSPERUNIT - 1; j++) {
                 mapLevel[posY + j][posX] = 1;
             }
 
-            */
         }
     }
